+++ docs-website/src/pages/cloud/index.js
@@ -9,12 +9,11 @@ import styles from "./styles.module.scss";
 import UnifiedTabs from "./UnifiedTabs";
 import FeatureCards from "./FeatureCards";
 import Hero from "./Hero";
+import DemoForm from "./DemoForm";
 
 function Home() {
   const context = useDocusaurusContext();
   const { siteConfig = {} } = context;
-  // const { colorMode } = useColorMode();
-
 
   if (siteConfig.customFields.isSaas) {
     window.location.replace("/docs");
@@ -37,25 +36,25 @@ function Home() {
       </div>
       <div className={clsx("hero", styles.hero)}>
         <div className="container" style={{ paddingTop: '12vh', paddingBottom: '12vh' }}>
-          <div className="hero__content">
-            <div>
-              <h1 className="hero__title">Get your free trial.</h1>
-              <div className={clsx(styles.hero__secondtitle)}>Data Discovery, Data Quality and Data Governance unified.</div>
-
-              <Link className="button button--primary button--lg" to="https://www.acryldata.io/datahub-sign-up?utm_source=datahub&utm_medium=referral&utm_campaign=acryl_signup">
-                Book Demo
-              </Link>
-              <Link className={clsx(styles.buttonLightBlue, "button button--secondary button--lg")} to="https://www.acryldata.io/tour">
-                Product Tour
-              </Link>
+          <div className={clsx(styles.hero__cta)}>
+          <div className="row row__padded">
+            <div  className={clsx(styles.col, "col col--7")}>
+              <h1 className={styles.hero__title}>Get your free trial.</h1>
+              <div className={clsx(styles.hero__subtitle)}>Data Discovery, Data Quality and Data Governance unified.</div>
+              <div>
+                <Link className={clsx(styles.button, styles.bookButton, "button button--primary button--lg")} to="https://www.acryldata.io/datahub-sign-up?utm_source=datahub&utm_medium=referral&utm_campaign=acryl_signup">
+                  Book Demo
+                </Link>
+                <Link className={clsx(styles.button, styles.productTourButton, "button button--secondary button--lg")} to="https://www.acryldata.io/tour">
+                  Live Product Tour →
+                </Link>
+              </div>
               <div className="hero__subtitle" />
-              {/* <hr style={{margin: "3rem"}}/>
-              <div className="hero__subtitle">
-                An extension of the DataHub Core project.<br/>
-                <a className={clsx(styles.link)} href="/docs/managed-datahub/managed-datahub-overview">View Cloud Docs.
-                </a>
-              </div> */}
             </div>
+            <div className={clsx(styles.col, "col col--5")}>
+               <DemoForm />
+            </div>
+          </div>
           </div>
         </div>
       </div>
